+++ example/lib/app.dart
@@ -1,6 +1,7 @@
-import 'languages/languages.dart';
+import 'languages/language_dart.g.dart';
+import 'widgets/widget_dart.g.dart';
 import 'package:finch/model.dart';
-import '../core/local_events.dart';
+import 'core/local_events.dart';
 import 'package:finch/finch_route.dart';
 import 'db/job_collection_free.dart';
 import 'db/person_collection_free.dart';
@@ -15,11 +16,12 @@ import 'route/web_route.dart';
 import 'package:finch/finch_capp.dart';
 
 FinchConfigs configs = FinchConfigs(
+  jinjaMapTemplate: mapTemplates,
   widgetsPath: pathTo(env['WIDGETS_PATH'] ?? "./lib/widgets"),
   widgetsType: env['WIDGETS_TYPE'] ?? 'j2.html',
   languagePath: pathTo(env['LANGUAGE_PATH'] ?? "./lib/languages"),
   languageSource: LanguageSource.dart,
-  dartLanguages: languages,
+  dartLanguages: languageDart,
   publicDir: pathTo(env['PUBLIC_DIR'] ?? './public'),
   dbConfig: FinchDBConfig(
     enable: true, //env['ENABLE_DATABASE'] == 'true',