+++ example/lib/middleware/test_middleware.dart
@@ -1,0 +1,9 @@
+import 'package:finch/finch_route.dart';
+
+class TestMiddleware extends Middleware {
+  @override
+  Future<String?> handle() async {
+    rq.addParam('middleware', "Test Middleware Active");
+    return null;
+  }
+}