+++ lib/src/cli/commands/commands.dart
@@ -1,12 +1,26 @@
 import 'dart:io';
 
 import 'package:capp/capp.dart';
+import 'package:finch/src/tools/convertor/language_to_dart.dart';
+import 'package:finch/src/tools/convertor/widget_to_dart.dart';
 import 'package:finch/src/tools/extensions/directory.dart';
 import 'package:finch/src/tools/path.dart';
 import 'package:finch/finch_app.dart';
 import 'package:archive/archive_io.dart';
+import 'package:yaml/yaml.dart';
+import 'package:path/path.dart' as p;
 
 class ProjectCommands {
+  Map finchConfigs = {};
+  ProjectCommands() {
+    var pubspecPath = _findPubspecPath(Directory.current.path);
+    var pubspec = _loadPubspec(pubspecPath);
+    var finchYaml = pubspec['finch'] as YamlMap?;
+    for (var key in finchYaml?.keys ?? []) {
+      finchConfigs[key] = finchYaml?[key];
+    }
+  }
+
   Future<CappConsole> get(CappController controller) async {
     await Process.start(
       'dart',
@@ -25,15 +39,23 @@ class ProjectCommands {
     return CappConsole('dart run build_runner build', CappColors.none);
   }
 
-  Future<CappConsole> run(CappController controller) async {
+  Future<CappConsole> run(
+    CappController controller, {
+    bool serve = false,
+  }) async {
     var path = controller.getOption('path');
+
     var defaultPath = [
       './bin',
       './lib',
       './src',
+      './example/bin',
+      './example/lib',
+      './example/src',
     ];
 
     var defaultApp = [
+      if (serve) 'serve.dart',
       'app.dart',
       'server.dart',
       'dart.dart',
@@ -41,6 +63,21 @@ class ProjectCommands {
       'run.dart',
       'watcher.dart',
     ];
+
+    if (path.isEmpty) {
+      var pubspecPath = _findPubspecPath(Directory.current.path);
+      var pubspec = _loadPubspec(pubspecPath);
+      if (pubspec.containsKey('finch')) {
+        var finchConfig = pubspec['finch'];
+        var appPathKey = serve ? 'serve' : 'app';
+        if (finchConfig[appPathKey] != null) {
+          var appPath = finchConfig[appPathKey];
+          if (appPath is String && appPath.isNotEmpty) {
+            path = appPath;
+          }
+        }
+      }
+    }
 
     if (path.isEmpty) {
       for (var p in defaultPath) {
@@ -65,16 +102,26 @@ class ProjectCommands {
       print("Running project from: $path");
     }
 
+    path = p.absolute(path);
     var proccess = await Process.start(
       'dart',
       [
         'run',
         "--enable-asserts",
         path,
+        '-v',
       ],
-      mode: ProcessStartMode.inheritStdio,
+      mode: ProcessStartMode.normal,
       workingDirectory: File(path).parent.parent.path,
     );
+
+    // Forward stdout and stderr to console
+    proccess.stdout.listen((data) {
+      stdout.add(data);
+    });
+    proccess.stderr.listen((data) {
+      stderr.add(data);
+    });
 
     var help = "Project is running (${proccess.pid})...\n\n"
         "┌┬┬┬┬┬┬┬┬┬┬┬┬┬┬┬┬┬┬┬┬───────────┬┬┬┬┬┬┬┬┬┬┬┬┬┬┬┬┬┬┬┬┐\n"
@@ -103,9 +150,18 @@ class ProjectCommands {
             'run',
             "--enable-asserts",
             path,
+            '-v',
           ],
-          mode: ProcessStartMode.inheritStdio,
+          mode: ProcessStartMode.normal,
         );
+
+        // Forward stdout and stderr to console
+        proccess.stdout.listen((data) {
+          stdout.add(data);
+        });
+        proccess.stderr.listen((data) {
+          stderr.add(data);
+        });
       } else if (['q', 'qy', 'qq'].contains(userInput.toLowerCase())) {
         var res = true;
         if (userInput.toLowerCase() == 'q') {
@@ -121,11 +177,12 @@ class ProjectCommands {
         CappConsole.write("Finch version: v${FinchApp.info.version}");
         CappConsole.write("Dart version: v${Platform.version}");
       } else {
-        CappConsole.write(
-          "Unknown input: ${userInput.toLowerCase()}",
-          CappColors.error,
-        );
-        CappConsole.write(help, CappColors.success);
+        try {
+          proccess.stdin.add(input);
+        } catch (e) {
+          CappConsole.write(
+              "Error sending input to process: $e", CappColors.error);
+        }
       }
     });
 
@@ -149,21 +206,32 @@ class ProjectCommands {
     return CappConsole("", CappColors.off);
   }
 
-  Future<CappConsole> build(CappController controller) async {
+  Future<CappConsole> build(
+    CappController controller, {
+    bool copyLang = true,
+    bool copyWidgets = true,
+  }) async {
     if (controller.existsOption('h')) {
       var help = controller.manager.getHelp([controller]);
       return CappConsole(help, CappColors.none);
     }
 
-    var path = controller.getOption('appPath', def: './lib/app.dart');
+    var path = controller.getOption(
+      'appPath',
+      def: finchConfigs['path'] ?? './lib/app.dart',
+    );
     if (path.isEmpty || !File(path).existsSync()) {
       return CappConsole(
           "The path of main file dart is requirment. for example '--path ./bin/app.dart'",
           CappColors.error);
     }
 
-    var output = controller.getOption('output', def: './finch_build');
-    if (output == './finch_build' && Directory(output).existsSync()) {
+    var defaultOutputPath = finchConfigs['build_output'] ?? './finch_build';
+    var output = controller.getOption(
+      'output',
+      def: defaultOutputPath,
+    );
+    if (output == defaultOutputPath && Directory(output).existsSync()) {
       Directory(output).deleteSync(recursive: true);
     } else if (Directory(output).existsSync()) {
       return CappConsole(
@@ -173,7 +241,11 @@ class ProjectCommands {
     }
     Directory(output).createSync(recursive: true);
 
-    var publicPath = controller.getOption('publicPath', def: './public');
+    var publicPath = controller.getOption(
+      'publicPath',
+      def: finchConfigs['public_path'] ?? './public',
+    );
+
     if (publicPath.isNotEmpty && Directory(publicPath).existsSync()) {
       var publicOutPutPath = joinPaths([output, 'public']);
       Directory(publicOutPutPath).createSync(recursive: true);
@@ -185,27 +257,51 @@ class ProjectCommands {
     }
 
     Directory('$output/lib').createSync(recursive: true);
+    var langPath = controller.getOption(
+      'langPath',
+      def: finchConfigs['languages_path'] ?? './lib/languages',
+    );
 
-    var langPath = controller.getOption('langPath', def: './lib/languages');
-    if (langPath.isNotEmpty && Directory(langPath).existsSync()) {
-      var langOutPutPath = joinPaths([output, 'lib/languages']);
-      Directory(langOutPutPath).createSync(recursive: true);
-      await CappConsole.progress(
-        "Copy Language files",
-        () => Directory(langPath).copyDirectory(Directory(langOutPutPath)),
-        type: CappProgressType.circle,
-      );
+    if (copyLang) {
+      if (langPath.isNotEmpty && Directory(langPath).existsSync()) {
+        var langOutPutPath = joinPaths([output, 'lib/languages']);
+        Directory(langOutPutPath).createSync(recursive: true);
+        await CappConsole.progress(
+          "Copy Language files",
+          () => Directory(langPath).copyDirectory(Directory(langOutPutPath)),
+          type: CappProgressType.circle,
+        );
+      }
+    } else {
+      await LanguageToDart(
+        langPath,
+        fileExtention: controller.getOption(
+          'languages_type',
+          def: '.${finchConfigs['languages_type'] ?? 'json'}',
+        ),
+      ).generate();
     }
 
-    var widgetPath = controller.getOption('widgetPath', def: './lib/widgets');
-    if (widgetPath.isNotEmpty && Directory(widgetPath).existsSync()) {
-      var widgetOutPutPath = joinPaths([output, 'lib/widgets']);
-      Directory(widgetOutPutPath).createSync(recursive: true);
-      await CappConsole.progress(
-        "Copy widgets",
-        () => Directory(widgetPath).copyDirectory(Directory(widgetOutPutPath)),
-        type: CappProgressType.circle,
-      );
+    var widgetPath = controller.getOption(
+      'widgetPath',
+      def: finchConfigs['widgets_path'] ?? './lib/widgets',
+    );
+    if (copyWidgets) {
+      if (widgetPath.isNotEmpty && Directory(widgetPath).existsSync()) {
+        var widgetOutPutPath = joinPaths([output, 'lib/widgets']);
+        Directory(widgetOutPutPath).createSync(recursive: true);
+        await CappConsole.progress(
+          "Copy widgets",
+          () =>
+              Directory(widgetPath).copyDirectory(Directory(widgetOutPutPath)),
+          type: CappProgressType.circle,
+        );
+      }
+    } else {
+      await WidgetToDart(
+        widgetPath,
+        fileExtention: '.${finchConfigs['widgets_type'] ?? 'html'}',
+      ).generate();
     }
 
     var envPath = controller.getOption('envPath', def: './.env');
@@ -265,5 +361,20 @@ class ProjectCommands {
 
     return CappConsole(
         'Finish build ${result == 0 ? 'OK!' : ''}', CappColors.none);
+  }
+
+  String _findPubspecPath(String startPath) {
+    var pubspecFile = File(joinPaths([startPath, 'pubspec.yaml']));
+    if (pubspecFile.existsSync()) {
+      return pubspecFile.path;
+    }
+    throw Exception('pubspec.yaml not found');
+  }
+
+  YamlMap _loadPubspec(String path) {
+    var pubspecFile = File(path);
+    var content = pubspecFile.readAsStringSync();
+    var pubspec = loadYaml(content);
+    return pubspec;
   }
 }