+++ lib/src/cli/commands/main.dart
@@ -2,6 +2,7 @@ import 'dart:io';
 
 import 'package:capp/capp.dart';
 import 'package:finch/src/finch_app.dart';
+import 'package:finch/src/tools/extensions/capp.dart';
 
 class Main {
   Future<CappConsole> main(CappController controller) async {
@@ -21,9 +22,6 @@ class Main {
       return CappConsole("Update Finch");
     }
 
-    return CappConsole(
-      controller.manager.getHelp(),
-      controller.existsOption('help') ? CappColors.none : CappColors.warning,
-    );
+    return controller.manager.writeHelpModern();
   }
 }