+++ lib/src/router/middleware.dart
@@ -1,0 +1,9 @@
+import 'package:finch/route.dart';
+
+abstract class Middleware {
+  Request get rq => Context.rq;
+
+  Middleware();
+
+  Future<String?> handle();
+}