+++ test/finch_server_test.dart
@@ -14,7 +14,8 @@ import 'package:html/parser.dart';
 void main() async {
   FinchApp server = FinchApp(
     configs: FinchConfigs(
-      port: 8089,
+      port: 0,
+      noStop: false,
       publicDir: 'public',
       languagePath: joinPaths([pathApp, '../example/lib/languages']),
       widgetsPath: '../example/lib/widgets',
@@ -258,6 +259,10 @@ void main() async {
   var httpServer = await server.start([], false).then((value) {
     Console.p("Example app started: http://localhost:${value.port}");
     return value;
+  });
+
+  tearDownAll(() async {
+    await httpServer.close(force: true);
   });
 
   group("Finch Server Test", () {